fix and rename  lpush-install, update package.json
