+++ plugin.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <plugin xmlns="http://apache.org/cordova/ns/plugins/1.0"
   xmlns:android="http://schemas.android.com/apk/res/android"
-  id="cordova-plugin-leanpush" version="0.1.2">
+  id="cordova-plugin-leanpush" version="0.1.3">
   <name>Cordova LeanCloud Push</name>
   <js-module src="www/LeanPush.js" name="LeanPush">
     <clobbers target="LeanPush"/>
